Project dockerized, File structure redesigned, Database changed,
