+++ Dockerfile
@@ -1,0 +1,7 @@
+FROM python:3.9
+WORKDIR /app
+COPY requirements.txt /app/requirements.txt
+RUN pip install -r requirements.txt
+COPY . /app
+
+CMD python run.py
