+++ main_app/templates/index.html
@@ -26,36 +26,36 @@
                 <label class="btn btn-outline-primary" for="btnradio1">
                     <a href=" {{ url_for('main.index') }} " class="nav-link active-nav-link"> Home </a>
                 </label>
-
+                 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio2" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio2">
-                    <a href=" {{ url_for('users.login') }} " class="nav-link"> Login  </a>
+                    <a href="#" class="nav-link"> Login  </a>
                 </label>
-
+                
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio3" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio3">
-                    <a href=" {{ url_for('users.login_parents') }} " class="nav-link">
+                    <a href=" {{ url_for('parent.login_parents') }} " class="nav-link">
                         Parents Login
                     </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio4" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio4">
-                    <a href=" {{ url_for('users.admission') }} " class="nav-link">
+                    <a href=" {{ url_for('student.admission') }} " class="nav-link">
                         Admission
                     </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio5" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio5">
-                    <a href=" {{ url_for('users.apply_parents') }} " class="nav-link">
+                    <a href=" {{ url_for('parent.apply_parents') }} " class="nav-link">
                         Parents Application Form
                     </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio6" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio6">
-                    <a href=" {{ url_for('users.apply_teacher') }} " class="nav-link">
+                    <a href=" {{ url_for('teacher.apply_teacher') }} " class="nav-link">
                         Teacher Application Form
                     </a>
                 </label>
@@ -67,10 +67,6 @@
                     </a>
                 </label>
             </div>
-
-            <!-- <h1> I will be right back! </h1>
-            <br />
-            <h2> Just 10 minutes </h2> -->
 
         <div>
             </div>
