+++ run.py
@@ -13,4 +13,4 @@ from main_app import create_app
 app = create_app()
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    app.run(host= '0.0.0.0', debug=True)
